#48 refactor : bindingAdapters 패키지 분리
